requirements.txt
selenium
PyYAML
capsolver_extension_python
retry